Fix CS1739 and CS0675 from step 7
- FlacChannelDecoderTests: use PascalCase named args (SampleRate,
  FrameOrSampleNumber, IsVariableBlockSize) to match record primary
  constructor parameter names.
- FlacFrameHeaderReader: cast cont & 0x3F to uint before OR-ing into
  long to avoid sign-extension warning (CS0675).

+++ src/Whirtle.Client/Codec/Flac/FlacFrameHeaderReader.cs
@@ -176,7 +176,7 @@ internal static class FlacFrameHeaderReader
             if ((cont & 0xC0) != 0x80)
                 throw new InvalidDataException(
                     $"Invalid UTF-8 continuation byte: 0x{cont:X2}.");
-            value = (value << 6) | (cont & 0x3F);
+            value = (value << 6) | (uint)(cont & 0x3F);
         }
 
         return value;
